refactor(max-score-finder): 探索結果表示を SearchResults に抽出
